Restructure into a workspace with PTX as a direct dependency
Move the PTX-backed implementations from ext/ into src/ptx/ and make PTX an
ordinary dependency; CUDACoreExt remains the only extension. Turn test/ and
docs/ into workspace members and split tests into host/, gpu/ and tools/
run through ParallelTestRunner, with TEST_TARGET banners gating runtime
sections per device capability. Add FP8 element types via Microfloats and
the shared test setup, targets and wgmma definitions.

+++ test/host/layouts.jl
@@ -91,7 +91,7 @@ end
     @test_throws DimensionMismatch Fragment((0f0,0f0),operand_layout(atom,Accumulator()))
     @test_throws ArgumentError MMAAtom((16,8,32),BFloat16)
     @test_throws ArgumentError MMAAtom((16,8,16),Float32)
-    @test_throws ArgumentError MMAAtom((16,8,16),Float16,Float16)
+    @test_throws ArgumentError MMAAtom((16,8,16),BFloat16,Float16)
     @test MMAAtom((16,8,8),Float32) isa MMAAtom
     p = TiledMMA(atom,Val((2,2)),Val((2,4)),Val(32))
     @test size(p) == (64,64,32)
